+++ changelog.md
@@ -3,6 +3,7 @@
 ## What's Changed
 
 - Add backend detection
+- Use CentralConnectionState in AioJsonRpcAioHttpClient
 
 # Version 2025.11.26 (2025-11-25)